MY PROJECT/PROJECT5 STREAMLIT/requirements.txt
cycler==0.10.0
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
Pillow==9.0.0
plotly==5.5.0
scikit_learn==1.0.2
seaborn==0.11.2
streamlit==1.4.0
